+++ sql/sql_plugin.h
@@ -172,7 +172,7 @@ extern plugin_ref plugin_lock(THD *thd, plugin_ref ptr);
 extern plugin_ref plugin_lock_by_name(THD *thd, const LEX_CSTRING *name,
                                       int type);
 extern void plugin_unlock(THD *thd, plugin_ref plugin);
-extern void plugin_unlock_list(THD *thd, plugin_ref *list, uint count);
+extern void plugin_unlock_list(THD *thd, plugin_ref *list, size_t count);
 extern bool mysql_install_plugin(THD *thd, const LEX_CSTRING *name,
                                  const LEX_CSTRING *dl);
 extern bool mysql_uninstall_plugin(THD *thd, const LEX_CSTRING *name,
